+++ src/utils/FieldSet.jsx
@@ -1,12 +1,13 @@
 import React from 'react'
 
 import BooleanField from '../fields/BooleanField'
-import { block, normalizeChoices } from '../utils'
+import { block } from '../utils'
+import choicesAdapter from './choicesAdapter'
 
+@choicesAdapter
 @block
 export default class FieldSet extends React.PureComponent {
   render(b) {
-    const normalizedChoices = normalizeChoices(this.props)
     const {
       type,
       isChecked,
@@ -19,7 +20,7 @@ export default class FieldSet extends React.PureComponent {
     } = this.props
     return (
       <fieldset className={b.mix(className)} ref={elementRef}>
-        {normalizedChoices.map(([choiceLabel, choice]) => (
+        {choices.map(([choiceLabel, choice]) => (
           <label
             key={choice}
             className={b.e('label').m({ on: isChecked(choice, value) })}