+++ src/utils/index.js
@@ -28,22 +28,3 @@ export const moneyFormat = price =>
   )
 
 export const block = blockMaker({ namespace: 'Formol_' })
-
-export const normalizeChoices = ({ choices }) =>
-  Array.isArray(choices)
-    ? choices.map(c => [c, c])
-    : Object.entries(choices || {})
-
-export const normalizeMultipleProps = ({ value, multiple, ...props }) => {
-  if (!multiple && Array.isArray(value)) {
-    value = value.length ? value[0] : null
-  }
-  if (multiple && !Array.isArray(value)) {
-    value = value !== null && value !== void 0 && value !== '' ? [value] : []
-  }
-  return {
-    value,
-    multiple,
-    ...props,
-  }
-}