+++ stories/utils.js
@@ -1,5 +1,5 @@
 import { withState } from '@dump247/storybook-state'
-import { boolean, number, select } from 'addon-knobs-null-number-fix'
+import { boolean, number, select } from '@storybook/addon-knobs'
 
 import Formol from '../src'
 import { diff } from '../src/utils/object'
@@ -15,6 +15,7 @@ export const withStateForm = (form, initial) =>
       },
       onChange: transient => store.set({ transient }),
       item: store.state.item,
+      transient: store.state.transient,
       readOnly: boolean('Form read only', false, 'form'),
       focusNextOnEnter: boolean('Focus next field with [Enter]', false, 'form'),
       i18n: select('I18n', Object.keys(Formol.i18n), 'en', 'form'),