+++ CHANGELOG.md
@@ -1,3 +1,30 @@
+### [1.2.4](https://github.com/Kozea/formol/compare/v1.2.3...v1.2.4)
+
+- Avoid crash on non-file files in FileField
+- Fix default formatter/normalizer
+
+### [1.2.3](https://github.com/Kozea/formol/compare/v1.2.2...v1.2.3)
+
+- Make SelectField super fast (even with a lot of choices).
+  Thanks to [react-virtualized-select](https://github.com/bvaughn/react-virtualized-select) and
+  [react-select-fast-filter-options](https://github.com/bvaughn/react-select-fast-filter-options)
+
+### [1.2.2](https://github.com/Kozea/formol/compare/v1.2.1...v1.2.2)
+
+- Use an updater for Formol setState to avoid data loss
+
+### [1.2.1](https://github.com/Kozea/formol/compare/v1.2.0...v1.2.1)
+
+- Fix null value in Select and bad coercive equality for value.
+
+## [1.2.0](https://github.com/Kozea/formol/compare/v1.1.10...v1.2.0)
+
+- Handle non-string values in select field (using an internal memo)
+- Add Inliner component to allow several fields in same line
+- Reset form on submit when no item was given
+- Allow restoring not entered state for fields from Form. Restore this state after a successful submit or a cancel.
+- Align password field (take eye size in account)
+
 ### [1.1.10](https://github.com/Kozea/formol/compare/v1.1.9...v1.1.10)
 
 - Fix typo for datetime-local type.